% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_product.R
\name{a_product}
\alias{a_product}
\title{A very simple function just for show}
\usage{
a_product(x, check = FALSE)
}
\arguments{
\item{x}{numeric. Input vector}

\item{check}{boolean. Should we check if x is numeric?}
}
\description{
A very simple function just for show
}
\examples{
x <- rexp(5)
a_product(x)

}
